Dev Changes for the updater
git-svn-id: file:///Users/Shitiz/Downloads/lssvn/source/limesurvey_dev@7638 b72ed6b6-b9f8-46b5-92b4-906544132732

+++ admin/update/updater.php
@@ -257,7 +257,6 @@ function UpdateStep4()
     $http->GetRequestArguments("http://update.limesurvey.org/updates/download/{$updateinfo['downloadid']}",$arguments);
     $http->RestoreCookies($_SESSION['updatesession']);      
 
-    $updateinfo=false;
     $error=$http->Open($arguments);           
     $error=$http->SendRequest($arguments);
     $error=$http->ReadReplyHeaders($headers);
@@ -280,8 +279,14 @@ function UpdateStep4()
     foreach ($updateinfo['files'] as $afile)
     {
         if ($afile['type']=='D' && file_exists($rootdir.$afile['file']))
-        {
-            unlink($rootdir.$afile['file']);
+        {   
+            if (is_file($rootdir.$afile['file']))
+            {
+                unlink($rootdir.$afile['file']);
+            }
+            else{
+                rmdirr($rootdir.$afile['file']);
+            }
             $output.=sprintf($clang->gT('File deleted: %s'),$afile['file']).'<br />'; 
         }
     }
